extract dpkg scanner from os scanner
